move supported asset type definitions to config.js
