+++ src/config.js
@@ -18,3 +18,20 @@ export const COLOR_PALETTE_BY_ENTITY = {
   node: 'node-color-seq',
   edge: 'edge-color-seq',
 };
+
+// Maps for supported asset types within the app. Used by asset chooser.
+// Remember to also update getNetworkType when changing these!
+
+export const SUPPORTED_MIME_TYPE_MAP = {
+  network: ['application/json', 'text/csv', 'application/vnd.ms-excel'],
+  image: ['image/*'],
+  audio: ['audio/*'],
+  video: ['video/*'],
+};
+
+export const SUPPORTED_EXTENSION_TYPE_MAP = {
+  network: ['.csv', '.json'],
+  image: ['.jpg', '.jpeg', '.gif', '.png'],
+  audio: ['.mp3', '.aiff'],
+  video: ['.mov', '.mp4'],
+};
